uni-ctf-2023/web/[Hard] PhantomFeed/challenge/phantom-market-backend/requirements.txt
cryptography==41.0.4
Flask==3.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
bcrypt==4.0.1 
reportlab==3.6.12
SQLAlchemy==2.0.21
